violet: Make bootimage with header version as 2
Adding support to make a bootimage with header version as 2.
This bootimage no longer has the dtbs appended to the kernel.
All the dtbs are made into a new image called dtb.img and
this image is present in the bootimage.

Do not append dtb to kernel image we are using retrofit dynamic
partition on violet.

Change-Id: I668bbcc3b3dce735c99aa41e843daaf7441697ff
Signed-off-by: Joker-V2 <nipinna0@gmail.com>